Add unit tests for Notifier service
